Clean up config file parsing
* Move config classes out of Component.cs
* Add JsonConverter to handle allowing both "game" and "name" as the key
  for the game name.
