+++ Config/Category.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+
+namespace LiveSplit.UI.Components
+{
+    internal class Category
+    {
+        public string name { get; set; }
+        public List<string> splits { get; set; }
+    }
+}
