+++ Config/Split.cs
@@ -1,0 +1,64 @@
+using System;
+using System.Collections.Generic;
+
+namespace LiveSplit.UI.Components
+{
+    internal class Split
+    {
+        public string name { get; set; }
+        public string address { get; set; }
+        public string value { get; set; }
+        public string type { get; set; }
+        public List<Split> more { get; set; }
+        public List<Split> next { get; set; }
+        public int posToCheck { get; set; } = 0;
+
+        public uint addressint { get { return Convert.ToUInt32(address, 16); } }
+        public uint valueint { get { return Convert.ToUInt32(value, 16); } }
+
+        public bool check(uint value, uint word)
+        {
+            bool ret = false;
+            switch (this.type)
+            {
+                case "bit":
+                    if ((value & this.valueint) != 0) { ret = true; }
+                    break;
+                case "eq":
+                    if (value == this.valueint) { ret = true; }
+                    break;
+                case "gt":
+                    if (value > this.valueint) { ret = true; }
+                    break;
+                case "lt":
+                    if (value < this.valueint) { ret = true; }
+                    break;
+                case "gte":
+                    if (value >= this.valueint) { ret = true; }
+                    break;
+                case "lte":
+                    if (value <= this.valueint) { ret = true; }
+                    break;
+                case "wbit":
+                    if ((word & this.valueint) != 0) { ret = true; }
+                    break;
+                case "weq":
+                    if (word == this.valueint) { ret = true; }
+                    break;
+                case "wgt":
+                    if (word > this.valueint) { ret = true; }
+                    break;
+                case "wlt":
+                    if (word < this.valueint) { ret = true; }
+                    break;
+                case "wgte":
+                    if (word >= this.valueint) { ret = true; }
+                    break;
+                case "wlte":
+                    if (word <= this.valueint) { ret = true; }
+                    break;
+            }
+            return ret;
+        }
+    }
+}
