+++ LiveSplit.USB2SNESSplitter.csproj
@@ -63,6 +63,11 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Config\Autostart.cs" />
+    <Compile Include="Config\Category.cs" />
+    <Compile Include="Config\Game.cs" />
+    <Compile Include="Config\ConfigFileJsonConverter.cs" />
+    <Compile Include="Config\Split.cs" />
     <Compile Include="USB2SnesW.cs" />
     <Compile Include="Component.cs" />
     <Compile Include="ComponentSettings.cs">
